refactor: change type to enum
