+++ src/data/hooks/useActiveAsset.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query"
 import { getActiveAsset as getActiveAsset_AAVE_V2_STABLE_CELLAR } from "data/actions/AAVE_V2_STABLE_CELLAR/getActiveAsset"
-import { ConfigProps } from "data/cellarDataMap"
+import { CellarKey, ConfigProps } from "data/types"
 import { AaveV2CellarV2 } from "src/abi/types"
 import { useCreateContracts } from "./useCreateContracts"
 
@@ -8,14 +8,14 @@ export const useActiveAsset = (config: ConfigProps) => {
   const { cellarContract } = useCreateContracts(config)
 
   const AAVE_V2_STABLE_CELLAR_QUERY_ENABLED = Boolean(
-    config.cellar.key === "AAVE_V2_STABLE_CELLAR" &&
+    config.cellar.key === CellarKey.AAVE_V2_STABLE_CELLAR &&
       cellarContract.provider
   )
 
   const query = useQuery(
     ["USE_ACTIVE_ASSET"],
     async () => {
-      if (config.cellar.key === "AAVE_V2_STABLE_CELLAR") {
+      if (config.cellar.key === CellarKey.AAVE_V2_STABLE_CELLAR) {
         return await getActiveAsset_AAVE_V2_STABLE_CELLAR(
           cellarContract as AaveV2CellarV2
         )
