+++ src/data/hooks/useApy.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query"
 import { getApy as getApy_AAVE_V2_STABLE_CELLAR } from "data/actions/AAVE_V2_STABLE_CELLAR/getApy"
-import { ConfigProps } from "data/cellarDataMap"
+import { CellarKey, ConfigProps } from "data/types"
 import { AaveV2CellarV2, SommStaking } from "src/abi/types"
 import { useCreateContracts } from "./useCreateContracts"
 
@@ -9,7 +9,7 @@ export const useApy = (config: ConfigProps) => {
     useCreateContracts(config)
 
   const AAVE_V2_STABLE_CELLAR_QUERY_ENABLED = Boolean(
-    config.cellar.key === "AAVE_V2_STABLE_CELLAR" &&
+    config.cellar.key === CellarKey.AAVE_V2_STABLE_CELLAR &&
       cellarContract.provider &&
       stakerContract?.provider
   )
@@ -17,7 +17,7 @@ export const useApy = (config: ConfigProps) => {
   const query = useQuery(
     ["USE_APY"],
     async () => {
-      if (config.cellar.key === "AAVE_V2_STABLE_CELLAR") {
+      if (config.cellar.key === CellarKey.AAVE_V2_STABLE_CELLAR) {
         return await getApy_AAVE_V2_STABLE_CELLAR(
           cellarContract as AaveV2CellarV2,
           stakerContract as SommStaking
