+++ src/data/hooks/useCurrentDeposits.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query"
 import { getCurrentDeposits as getCurrentDeposits_AAVE_V2_STABLE_CELLAR } from "data/actions/AAVE_V2_STABLE_CELLAR/getCurrentDeposits"
-import { ConfigProps } from "data/cellarDataMap"
+import { CellarKey, ConfigProps } from "data/types"
 import { useGetCellarQuery } from "generated/subgraph"
 
 export const useCurrentDeposits = (config: ConfigProps) => {
@@ -16,7 +16,7 @@ export const useCurrentDeposits = (config: ConfigProps) => {
     cellar || {}
 
   const AAVE_V2_STABLE_CELLAR_QUERY_ENABLED = Boolean(
-    config.cellar.key === "AAVE_V2_STABLE_CELLAR" &&
+    config.cellar.key === CellarKey.AAVE_V2_STABLE_CELLAR &&
       asset?.symbol &&
       addedLiquidityAllTime &&
       removedLiquidityAllTime
@@ -25,7 +25,7 @@ export const useCurrentDeposits = (config: ConfigProps) => {
   const query = useQuery(
     ["USE_CURRENT_DEPOSITS"],
     async () => {
-      if (config.cellar.key === "AAVE_V2_STABLE_CELLAR") {
+      if (config.cellar.key === CellarKey.AAVE_V2_STABLE_CELLAR) {
         return await getCurrentDeposits_AAVE_V2_STABLE_CELLAR({
           assetSymbol: asset?.symbol,
           addedLiquidityAllTime,
