+++ src/data/hooks/useUserStakes.tsx
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query"
 import { getUserStakes as getUserStakes_AAVE_V2_STABLE_CELLAR } from "data/actions/AAVE_V2_STABLE_CELLAR/getUserStakes"
-import { ConfigProps } from "data/cellarDataMap"
+import { ConfigProps, StakerKey } from "data/types"
 import { SommStaking } from "src/abi/types"
 import { useCreateContracts } from "./useCreateContracts"
 import { useAccount } from "wagmi"
@@ -11,7 +11,7 @@ export const useUserStakes = (config: ConfigProps) => {
 
   const AAVE_STAKER_QUERY_ENABLED = Boolean(
     address &&
-      config.staker?.key === "AAVE_STAKER" &&
+      config.staker?.key === StakerKey.AAVE_STAKER &&
       stakerContract?.provider &&
       stakerSigner?.provider &&
       stakerSigner.signer
@@ -20,7 +20,7 @@ export const useUserStakes = (config: ConfigProps) => {
   const query = useQuery(
     ["USE_USER_STAKES", address],
     async ({ queryKey: [, _address] }) => {
-      if (config.staker?.key === "AAVE_STAKER") {
+      if (config.staker?.key === StakerKey.AAVE_STAKER) {
         return await getUserStakes_AAVE_V2_STABLE_CELLAR(
           _address!,
           stakerContract as SommStaking,
