+++ src/data/types.ts
@@ -1,0 +1,61 @@
+// if there is a new abi, add the key here (ex: "AAVE_V2_STABLE_CELLAR" | "AAVE_V1_STABLE_CELLAR")
+
+export enum StakerKey {
+  AAVE_STAKER = "AAVE_STAKER",
+}
+export enum CellarKey {
+  AAVE_V2_STABLE_CELLAR = "AAVE_V2_STABLE_CELLAR",
+}
+export enum CellarRouterKey {
+  CELLAR_ROUTER = "CELLAR_ROUTER",
+}
+
+export interface ConfigProps {
+  id: string
+  lpToken: {
+    address: string
+  }
+  cellarRouter: {
+    address: string
+    abi: unknown
+    key: CellarRouterKey
+  }
+  cellar: {
+    address: string
+    abi: unknown
+    key: CellarKey
+  }
+  // staker optional because there will be a cellar without staker contract
+  staker?: {
+    address: string
+    abi: unknown
+    key: StakerKey
+  }
+  rewardTokenAddress: string
+}
+export interface CellarDataMap {
+  [key: string]: {
+    name: string
+    description: string
+    strategyType: string
+    managementFee: string
+    individualApy: number
+    cellarApy: number
+    protocols: string
+    supportedChains: string[]
+    performanceSplit: {
+      [key: string]: number
+    }
+    strategyBreakdown: {
+      [key: string]: string
+    }
+    strategyImgSrc?: string
+    strategyProvider?: {
+      logo?: string
+      title?: string
+      href?: string
+      tooltip?: string
+    }
+    config: ConfigProps
+  }
+}
